% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrrgression-package.R
\docType{package}
\name{regressionworlddata-package}
\alias{regressionworlddata-package}
\alias{regressionworlddata}
\title{Moinput Regression function library}
\description{
Package contains functions to estimate model parameters
}
\details{
\tabular{ll}{ Package: \tab regressionworlddata\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2016-09-23\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Benjamin Leon Bodirsky, Antonia Walther

Maintainer:  Benjamin Leon Bodirsky <bodirsky@pik-potsdam.de>
}
