% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_vce.R
\name{robust_vce}
\alias{robust_vce}
\title{robust_vce}
\usage{
robust_vce(x)
}
\arguments{
\item{x}{regression model}
}
\value{
a robust estimte of variance-covariance matrix and corresponding t-value and p-value for estimated coefficients
}
\description{
returns robust var-cov estimate
}
\author{
Xiaoxi Wang
}
