% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCollectRegressionVariables.R
\name{toolCollectRegressionVariables}
\alias{toolCollectRegressionVariables}
\title{toolCollectRegressionVariables}
\usage{
toolCollectRegressionVariables(indicators)
}
\arguments{
\item{indicators}{todo}
}
\value{
todo
}
\description{
todo
}
\author{
Benjamin Leon Bodirsky
}
