% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRegression.R
\name{toolRegression}
\alias{toolRegression}
\title{toolRegression}
\usage{
toolRegression(
  denominator,
  quotient = NULL,
  func = y ~ (a * x)/(b + x),
  x = "IHME_USD05_PPP_pc",
  z = NULL,
  ylab = NULL,
  xlab = NULL,
  data = NULL,
  countries_nlsAddLines = NULL,
  weight = "pop",
  x_log10 = FALSE,
  crossvalid_sample = NULL,
  crossvalid_drawing = 1,
  ...
)
}
\arguments{
\item{denominator}{denominator of the dependent variable that shall be estimated using the regression}

\item{quotient}{quotient of the dependent variable that shall be estimated using the regression}

\item{func}{functional relation for the regression, shall be in the format y~f(x,...) with x being gdp, y being denominator/quotient, and f() being any type of functional relationship. ... can inlcude either z or parameters to be estimated.}

\item{x}{independet variable, by default income}

\item{z}{additional independet variable}

\item{ylab}{name of y axis}

\item{xlab}{name of x axis}

\item{data}{data can be provided if Data shall not be derived by mrcommons:::calcCollectFoodDemandRegressionData()}

\item{countries_nlsAddLines}{the number of weightiest countries or the name of countries that shall be plotted by lines in the plot}

\item{weight}{the weight}

\item{x_log10}{passed on to nlsregression()}

\item{crossvalid_sample}{sample name from madrat used for crossvalidation. Name is built as follows:
crossvalid_seedX_kY
X is the random seed,
Y is the number of drawings. The combination of all drawings is the full sample.}

\item{crossvalid_drawing}{selected drawing of k in crossvalidsample}

\item{...}{further attributes that will be handed on to nlsregression(): 

An additional explanatory variable z can be added.
A regression model has to be chosen.
Startvalues can be predetermained.}
}
\value{
regression plot and the parameters from nlsregression
}
\description{
Regression model for the correlation of a denominator and quotient to the GDP, allowing for an additional driver z next to income.
}
\examples{

\dontrun{ 

toolRegression(denominator=livestock,
                     func=y~(a*x)/(b+x),
                     z=NULL,
                     startvalues=list(a=1100,b=7770)
                     )
                     
toolRegression(denominator=findset("kap"),
                     quotient=findset("kfo"),
                     func=y~(a*x)/(b+x),
                     z=NULL,
                     startvalues=list(a=0.5,b=7770)
                     )
                     
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Antonia Walther, Benjamin Leon Bodirsky
}
